# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 库存调拨规划器 API
  version: 1.0.0
  description: |
    库存调拨规划器提供完整的调拨申请、库存约束检查、调拨方案生成与确认、多仓自动分配、
    仓库管理、SKU 管理、库存管理及调拨记录查询等能力。

    **主要模块：**
    - 仓库管理 (`/api/warehouses`)
    - SKU 管理 (`/api/skus`)
    - 库存管理 (`/api/inventory`)
    - 调拨申请 (`/api/transfer-requests`)
    - 调拨方案 (`/api/transfer-plans`)
    - 调拨记录 (`/api/transfer-records`)

    **约束分级：**
    - `FATAL`：不可通过，阻断调拨
    - `WARNING`：可确认但需业务复核
    - `INFO`：提示性信息

    **调拨方案状态机：**
    - `PENDING`（待确认）→ `CONFIRMED`（已确认）→ `COMPLETED`（已完成）
    - `PENDING` → `REJECTED`（已驳回）
    - 任意状态 → `CANCELLED`（已取消）
  contact:
    name: Inventory Ops Team
servers:
  - url: http://localhost:3000
    description: 本地开发环境
  - url: /
    description: 当前部署
...
paths:
...
  /api/warehouses:
    get:
      tags: [Warehouses]
      summary: 查询仓库列表
      operationId: listWarehouses
      parameters:
        - $ref: '#/components/parameters/StatusQuery'
        - name: type
          in: query
          schema: { type: string, enum: [CENTER, REGIONAL, NORMAL] }
          description: 仓库类型筛选
        - name: keyword
          in: query
          schema: { type: string, maxLength: 50 }
          description: 关键字搜索（名称/编码/位置）
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 仓库列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListResponse'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Warehouse' } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      tags: [Warehouses]
      summary: 创建仓库
      operationId: createWarehouse
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WarehouseCreate' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CreatedResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Warehouse' }
        '400':
          description: 编码重复或必填项缺失
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/warehouses/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Warehouses]
      summary: 查询单个仓库详情
      operationId: getWarehouse
      responses:
        '200':
          description: 仓库详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Warehouse' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Warehouses]
      summary: 更新仓库信息
      operationId: updateWarehouse
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WarehouseUpdate' }
      responses:
        '200':
          description: 更新成功
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Warehouses]
      summary: 删除仓库
      operationId: deleteWarehouse
      description: 仅当仓库下无库存记录时允许删除
      responses:
        '200': { description: 删除成功 }
        '400':
          description: 仓库下存在库存记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404': { $ref: '#/components/responses/NotFound' }
  /api/warehouses/{id}/inventory:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Warehouses]
      summary: 查询指定仓库的库存明细
      operationId: listWarehouseInventory
      parameters:
        - name: sku_id
          in: query
          schema: { type: integer, minimum: 1 }
        - name: low_stock
          in: query
          schema: { type: boolean }
          description: 仅显示低库存（available<=min_stock）
      responses:
        '200':
          description: 库存明细列表
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1, example: 1 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
...
    StatusQuery:
      name: status
      in: query
      schema: { type: string, enum: [ACTIVE, INACTIVE], default: ACTIVE }
  responses:
    ValidationError:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorBody'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    ServerError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
    Warehouse:
      type: object
      properties:
        id: { type: integer }
        code: { type: string, example: WH-BJ }
        name: { type: string, example: 北京中心仓 }
        location: { type: string }
        type: { type: string, enum: [CENTER, REGIONAL, NORMAL] }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
        min_security_stock: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WarehouseCreate:
      type: object
      required: [code, name]
      properties:
        code: { type: string, maxLength: 32 }
        name: { type: string, maxLength: 100 }
        location: { type: string, maxLength: 200 }
        type: { type: string, enum: [CENTER, REGIONAL, NORMAL], default: NORMAL }
        status: { type: string, enum: [ACTIVE, INACTIVE], default: ACTIVE }
        min_security_stock: { type: integer, minimum: 0, default: 0 }
    WarehouseUpdate:
      allOf:
        - $ref: '#/components/schemas/WarehouseCreate'
        - type: object
...
    ListResponse:
      type: object
      properties:
        data: { type: array }
        total: { type: integer, minimum: 0 }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }
    CreatedResponse:
      type: object
      properties:
        message: { type: string }
    ErrorBody:
      type: object
      required: [error, code]
      properties:
        error: { type: string, description: 面向用户的错误消息 }
        code:
          type: string
          description: 稳定的机器可识别错误码
          example: WAREHOUSE_NOT_FOUND
        requestId:
          type: string
          description: 请求追踪 ID，用于日志关联
        timestamp: { type: string, format: date-time }
    ValidationErrorBody:
      allOf:
        - $ref: '#/components/schemas/ErrorBody'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                required: [path, message]
                properties:
                  path: { type: string, example: target_warehouse_id }
                  message: { type: string, example: target_warehouse_id 必填 }
                  value: { type: any }